usersApi/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
pymysql
python-dotenv
flask-cors
flask-bcrypt
flask-jwt-extended
flask-login